requirements.txt
# python 3.12
openai-clip
transformers==4.40.0
huggingface-hub
matplotlib
open_clip_torch
peft==0.10.0
pillow
tqdm
setuptools
datasets
openpyxl
datasets
numpy
pandas
tokenizers
pycocotools
pycocoevalcap
trl
opencv-python
evaluate
sacrebleu
statsmodels
scikit-learn
rouge-score

